pyproject.toml
[project]
name = "community-oshwa-org"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.2",
    "jinja2>=3.1.4",
    "python-dateutil>=2.9.0.post0",
    "rich>=13.9.4",
]
